requirements.txt
streamlit>=1.30.0
pandas
numpy
scikit-learn
torch
transformers
playwright
streamlit-option-menu
google-generativeai
matplotlib
